Updated code from prince
